clients/android: Fix parsing of add obstacle requests (#163)

+++ clients/AlphabotAndroidClient/app/src/main/java/at/mg6/filip/alphabotandroidclient/LPSView.java
@@ -210,7 +210,7 @@ public class LPSView extends View implements View.OnTouchListener {
         return Collections.unmodifiableList(obstacles);
     }
 
-    public Obstacle getObstacle(short id) {
+    public Obstacle getObstacle(int id) {
         for (Obstacle obstacle : obstacles)
             if (obstacle.hasId() && obstacle.getId() == id)
                 return obstacle;
